Create a Subscription class
